Migrate more pushes to explicit immediates
This is a kind of grab-bag of changes to use methods that are explicitly
about immediates when we know for sure they are.

This buys us two things:

1. It avoids the need to check the value before storing it in the stack
   or loading it from the stack.
2. It pushes back the line of "code that works directly with JS
   objects". Even when dealing with values like JavaScript's undefined,
   operations like `dup` and `copy` simply copy the encoded immediate
   directly and never need to enter JS.

In general, avoiding JS values unnecessarily should make it easier to
isolate parts of the code to be used directly in wasm (or wasm-like
code).

This code still isolates encoding and decoding to loads and stores in
the stack, but it probably makes sense to allow opcodes to understand
immediates directly, if they are working with values they know are not
JavaScript references.